requirements.txt
ffprobe-python==1.0.3
pymkv==1.0.8
python-opensubtitles==0.6.dev0
requests==2.25.1
charset-normalizer==1.3.4
watchdog==1.0.2